+++ src/association.ts
@@ -1,5 +1,5 @@
 /** Contains the association types. */
-import { defineAssociation, hasModelOptions } from './metadata';
+import { hasModelOptions } from './metadata';
 import { Model, ModelConstructor } from './model';
 import { Property } from './property';
 
@@ -94,31 +94,4 @@ export function isLazyLoad<T extends Model>(target: AssociationTarget<T>): boole
   // even if it's some arbitrary value and not a lazy loader, the safe
   // will catch that because it requires a model decorator to be associated.
   return !hasModelOptions(target);
-}
-
-/**
- * A decorator for a model association.
- * The association has a type and the model to associate to.
- * All association properties must be defined using this decorator in order
- * to be recognised by ModelSafe.
- *
- * Here the target is optional incase it is yet to be defined.
- * Once it has been defined, it can be correctly reassociated with
- * a target model using `Model.assocate`.
- *
- * @see `Model.associate`
- * @param type    The type of association.
- * @param target  The target model to associate to.
- * @param options Any extra Sequelize attribute options required.
- */
-export function assoc<T extends Model>(type: AssociationType, target?: AssociationTarget<T>,
-                                       options?: AssociationOptions) {
-  return (ctor: object, key: string | symbol) => defineAssociation(ctor, key, {
-    readOnly: false,
-
-    ... options,
-
-    type,
-    target
-  });
 }